bin/requirements.txt
google.cloud
requests
slack_sdk
google-cloud-secret-manager
aiohttp